# apidoc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "2.0.0",
    "title": "Identifier Services API v2",
    "contact": {
      "name": "Identifier Services support",
      "url": "https://www.kansalliskirjasto.fi/fi/palvelut/isbn-tunnus"
    }
  },
...
  "paths": {
...
    "/v2/monograph/isbn-ranges": {
      "get": {
        "operationId": "getIsbnRanges",
        "summary": "Read all ISBN ranges information",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "tags": [
          "monograph",
          "isbn"
        ],
        "responses": {
          "200": {
            "description": "Resources could be read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/isbn-ranges"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      },
      "post": {
        "operationId": "createIsbnRange",
        "summary": "Create a new ISBN range",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "description": "Creates a new ISBN range. Note new range is not allowed to overlap any existing ranges.",
        "tags": [
          "monograph",
          "isbn"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/isbn-range-create-body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/created"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      }
    },
    "/v2/monograph/isbn-ranges/{id}": {
      "get": {
        "operationId": "readIsbnRange",
        "summary": "Read ISBN range information",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "tags": [
          "monograph",
          "isbn"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Resource was read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/isbn-range-read"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource was not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      },
      "patch": {
        "operationId": "updateIsbnRange",
        "summary": "Update ISBN range",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "tags": [
          "monograph",
          "isbn"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/isbn-range-update-body"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Resource was updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource was not found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      },
      "delete": {
        "operationId": "deleteIsbnRange",
        "summary": "Delete ISBN range",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "tags": [
          "monograph",
          "isbn"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Resource was deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource was not found"
          },
          "409": {
            "description": "Conflict"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      }
    },
    "/v2/monograph/isbn-publisher-ranges": {
      "post": {
        "operationId": "createIsbnPublisherRange",
        "summary": "Create a new ISBN publisher range",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "description": "Creates a new ISBN publisher range with associations to ISBN range and monograph publisher.",
        "tags": [
          "monograph",
          "isbn"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/isbn-publisher-range-create-body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/created"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      }
    },
    "/v2/monograph/isbn-publisher-ranges/{id}": {
      "delete": {
        "operationId": "deleteIsbnPublisherRange",
        "summary": "Delete ISBN publisher range",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "tags": [
          "monograph",
          "isbn"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Resource was deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource was not found"
          },
          "409": {
            "description": "Conflict"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      }
    },
    "/v2/monograph/ismn-ranges": {
      "get": {
        "operationId": "getIsmnRanges",
        "summary": "Read all ISMN ranges information",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "tags": [
          "monograph",
          "ismn"
        ],
        "responses": {
          "200": {
            "description": "Resources could be read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ismn-ranges"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      },
      "post": {
        "operationId": "createIsmnRange",
        "summary": "Create a new ISMN range",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "description": "Creates a new ISMN range. Note new range is not allowed to overlap any existing ranges.",
        "tags": [
          "monograph",
          "ismn"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ismn-range-create-body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/created"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      }
    },
    "/v2/monograph/ismn-ranges/{id}": {
      "get": {
        "operationId": "readIsmnRange",
        "summary": "Read ISMN range information",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "tags": [
          "monograph",
          "ismn"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Resource was read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ismn-range-read"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource was not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      },
      "patch": {
        "operationId": "updateIsmnRange",
        "summary": "Update ISMN range",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "tags": [
          "monograph",
          "ismn"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ismn-range-update-body"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Resource was updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource was not found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      },
      "delete": {
        "operationId": "deleteIsmnRange",
        "summary": "Delete ISMN range",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "tags": [
          "monograph",
          "ismn"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Resource was deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource was not found"
          },
          "409": {
            "description": "Conflict"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      }
    },
    "/v2/monograph/ismn-publisher-ranges": {
      "post": {
        "operationId": "createIsmnPublisherRange",
        "summary": "Create a new ISMN publisher range",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "description": "Creates a new ISMN publisher range with associations to ISMN range and monograph publisher.",
        "tags": [
          "monograph",
          "ismn"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ismn-publisher-range-create-body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource was created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/created"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      }
    },
    "/v2/monograph/ismn-publisher-ranges/{id}": {
      "delete": {
        "operationId": "deleteIsmnPublisherRange",
        "summary": "Delete ISMN publisher range",
        "security": [
          {
            "openId": [
              "admin"
            ]
          }
        ],
        "tags": [
          "monograph",
          "ismn"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "204": {
            "description": "Resource was deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Resource was not found"
          },
          "409": {
            "description": "Conflict"
          },
          "default": {
            "$ref": "#/components/responses/error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "openId": {
        "type": "openIdConnect"
      }
    },
    "schemas": {
      "error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "URI where user can read more about problem",
            "example": "about:blank"
          },
          "status": {
            "type": "number",
            "description": "HTTP status code",
            "example": 401
          },
          "title": {
            "type": "string",
            "description": "Error title",
            "example": "Unauthorized"
          },
          "detail": {
            "type": "string",
            "description": "Error description",
            "example": "The requested operation is not permitted for unauthorized users."
          },
          "instance": {
            "type": "string",
            "description": "Original request path",
            "example": "/monograph/isbn-ranges/1"
          }
        }
      },
...
      "created-metadata": {
        "type": "object",
        "properties": {
          "created_by": {
            "type": "string",
            "minLength": 1,
            "maxLength": 36,
            "example": "foobar"
          },
          "created": {
            "type": "string",
            "example": "2024-12-01T00:00:00.000Z"
          }
        }
      },
      "modified-metadata": {
        "type": "object",
        "properties": {
          "modified_by": {
            "type": "string",
            "minLength": 1,
            "maxLength": 36,
            "example": "foobar"
          },
          "modified": {
            "type": "string",
            "example": "2024-12-01T00:00:00.000Z"
          }
        }
      },
      "isbn-range-base": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "gs1": {
                "type": "string",
                "example": "979"
              },
              "registration_group": {
                "type": "string",
                "example": "952"
              },
              "range_begin": {
                "type": "string",
                "example": "0"
              },
              "range_end": {
                "type": "string",
                "example": "1"
              },
              "active": {
                "type": "boolean",
                "example": true
              }
            }
          },
          {
            "$ref": "#/components/schemas/created-metadata"
          },
          {
            "$ref": "#/components/schemas/modified-metadata"
          }
        ]
      },
      "isbn-ranges": {
        "type": "array",
        "items": {
          "allOf": [
            {
              "$ref": "#/components/schemas/isbn-range-base"
            },
            {
              "type": "object",
              "properties": {
                "taken": {
                  "type": "integer",
                  "example": 1
                },
                "free": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          ]
        }
      },
      "isbn-range-create-body": {
        "type": "object",
        "properties": {
          "gs1": {
            "type": "string",
            "enum": [
              "978",
              "979"
            ]
          },
          "registration_group": {
            "type": "string",
            "enum": [
              "951",
              "952"
            ]
          },
          "range_begin": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5,
            "example": "99000"
          },
          "range_end": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5,
            "example": "99999"
          }
        },
        "required": [
          "gs1",
          "registration_group",
          "range_begin",
          "range_end"
        ]
      },
      "created": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 123
          }
        }
      },
      "isbn-range-read": {
        "allOf": [
          {
            "$ref": "#/components/schemas/isbn-range-base"
          },
          {
            "type": "object",
            "properties": {
              "available_publisher_ranges": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "isbn-range-update-body": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "example": true
          },
          "range_begin": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5,
            "example": "99000"
          },
          "range_end": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5,
            "example": "99999"
          }
        }
      },
      "isbn-publisher-range-create-body": {
        "type": "object",
        "properties": {
          "publisher_identifier": {
            "type": "string"
          },
          "isbn_range_id": {
            "type": "number"
          },
          "monograph_publisher_id": {
            "type": "number"
          }
        },
        "required": [
          "publisher_identifier",
          "isbn_range_id",
          "monograph_publisher_id"
        ]
      },
      "ismn-range-base": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "gs1": {
                "type": "string",
                "example": "979"
              },
              "registration_group": {
                "type": "string",
                "example": "0"
              },
              "range_begin": {
                "type": "string",
                "example": "100"
              },
              "range_end": {
                "type": "string",
                "example": "101"
              },
              "active": {
                "type": "boolean",
                "example": true
              }
            }
          },
          {
            "$ref": "#/components/schemas/created-metadata"
          },
          {
            "$ref": "#/components/schemas/modified-metadata"
          }
        ]
      },
      "ismn-ranges": {
        "type": "array",
        "items": {
          "allOf": [
            {
              "$ref": "#/components/schemas/ismn-range-base"
            },
            {
              "type": "object",
              "properties": {
                "taken": {
                  "type": "integer",
                  "example": 1
                },
                "free": {
                  "type": "integer",
                  "example": 1
                }
              }
            }
          ]
        }
      },
      "ismn-range-create-body": {
        "type": "object",
        "properties": {
          "gs1": {
            "type": "string",
            "enum": [
              "979"
            ]
          },
          "registration_group": {
            "type": "string",
            "enum": [
              "0"
            ]
          },
          "range_begin": {
            "type": "string",
            "minLength": 3,
            "maxLength": 7,
            "example": "99000"
          },
          "range_end": {
            "type": "string",
            "minLength": 3,
            "maxLength": 7,
            "example": "99999"
          }
        },
        "required": [
          "gs1",
          "registration_group",
          "range_begin",
          "range_end"
        ]
      },
      "ismn-range-read": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ismn-range-base"
          },
          {
            "type": "object",
            "properties": {
              "available_publisher_ranges": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "ismn-range-update-body": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "example": true
          },
          "range_begin": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5,
            "example": "99000"
          },
          "range_end": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5,
            "example": "99999"
          }
        }
      },
      "ismn-publisher-range-create-body": {
        "type": "object",
        "properties": {
          "publisher_identifier": {
            "type": "string"
          },
          "ismn_range_id": {
            "type": "number"
          },
          "monograph_publisher_id": {
            "type": "number"
          }
        },
        "required": [
          "publisher_identifier",
          "ismn_range_id",
          "monograph_publisher_id"
        ]
      },
...
    },
    "responses": {
      "error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/error"
            }
          }
        }
      }
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "The id of the entity to retrieve or interact with",
        "schema": {
          "type": "integer",
          "example": 123
        }
      }
    }
  }
}